Revert tests changes
