--- src/test/groovy/org/gradle/profiler/IntellijIntegrationTest.groovy
@@ -1,13 +1,0 @@
-package org.gradle.profiler
-
-import org.gradle.profiler.studio.tools.IntellijFinder
-import spock.lang.Requires
-
-@Requires({ IntellijFinder.findIdeHome() })
-class IntellijIntegrationTest extends AbstractIdeSyncIntegrationTest {
-
-    @Override
-    File findIdeHome() {
-        return IntellijFinder.findIdeHome()
-    }
-}
